Move `Widgets.WidgetArgs` to its own file
